# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Mercury Express API
  version: 1.2.0
  description: API для управления новостями и профессиональным складским учетом (FIFO, Партии, Транзакции, Отчеты)
...
paths:
...
  /news:
    get:
      summary: Список всех новостей
      tags: [News]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/News" }
    post:
      summary: Создать новость с файлами
      tags: [News]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title: { type: string }
                content: { type: string }
                categoryId: { type: integer }
                files: { type: array, items: { type: string, format: binary } }
      responses:
        201: { description: Created }

  # --- ЗАПЧАСТИ ---
  /parts:
    get:
      summary: Все запчасти и их общие остатки
      tags: [Parts]
      responses:
        200: { description: OK }
    post:
      summary: Создать карточку запчасти
      tags: [Parts]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                categoryId: { type: integer }
                description: { type: string }
      responses:
        201: { description: Created }
...
components:
  schemas:
    News:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: string }
        Files:
          type: array
          items:
            type: object
            properties:
              url: { type: string }
...